Camera: Remove meta data stream from snapshot channel.
This is qcamera test app specific change where
we remove metadata stream from snapshot channel.

This is temporary chage to fix ANR in snapshot
with test app. In future, we can fix it with
one of the following ways

1. we can add thumbnail stream as well
to snapshot channel along with metadata.

2. Fix the issue on back end side by adding meta
data generation in buf_done on snapshot frames.

Change-Id: Ief690045a9157177d42dca2af030581fa8387fed