QCamera2: Work around passive autofocus stalls
When a photo is captured before passive AF finishes sometimes, passive AF
will stay stuck on scanning forever (without _actually_ scanning) until the
backend decides to start a new passive AF event or until the camera is
restarted. Normally this would be a minor issue that wouldn't be noticed,
but since the camera app now decides when to capture photos based on the
state of passive AF, this bug can completely deadlock the camera app until
one of two conditions mentioned above is fulfilled.

This is a bug in the camera backend, so it can't truly be fixed from the
HAL. In order to work around the bug, this patch will detect when passive
AF stalls and notify userspace that the passive AF event is over so that
the camera app doesn't deadlock.