QCamera2: Send a static set of camera parameters for flashlight init
For some reason (backend bugs?), camera_open() fails somewhat frequently
before the device is fully booted. The flashlight class initializes itself
by opening the cameras and retrieving their parameters before the device is
fully booted, so this creates a problem. When camera_open() fails during
the flashlight class' probe, the flashlight class will report no flash LEDs
found and thus the system flashlight will be broken.

In order to fix this, eliminate the need to call camera_open() at all when
the flashlight class attempts to open the camera. All the flashlight class
needs is a set of working camera parameters, so send it a static set of
actual camera parameters defined at compile time in order to make its
initialization successful 100% of the time.