QCamera2: Remove ISO-based antishake
Due to the unstable nature of how the ISO was being changed (i.e. there are
a bunch of hacks needed to change the ISO with our camera stack), this
feature is unstable and results in the camera preview becoming severely
underexposed after capturing photos sometimes.

The amount of motion blur reduction this provides is negligible (since
our chromatix library for the camera preview caps exposure at 1/30th of
a second), so just remove the feature entirely.